# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marasem API",
        "description": "This API supports social media login with Google, Facebook, and Behance. Users are redirected to the provider's login page and upon successful authentication, they are logged into the app and issued a token.",
        "contact": {
            "email": "amrfayez.247@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server for API routes"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local development server for Web routes"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Register a new user",
                "operationId": "6cb93c5db517a3e6e369c44ede7f22fd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "phone",
                                    "country_code"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8,
                                        "example": "password123"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "pattern": "^[0-9]{7,15}$",
                                        "example": "1234567890"
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "pattern": "^\\\\+\\\\d{1,4}$",
                                        "example": "+1"
                                    },
                                    "is_artist": {
                                        "description": "Optional. Indicates if the user is an artist.",
                                        "type": "boolean",
                                        "example": true,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User registered successfully."
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john.doe@example.com"
                                                },
                                                "is_artist": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "sample-jwt-token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/add-social-media-links": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Add social media links for the user",
                "operationId": "a070cfba95e6efb697a1069f08f780d9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "social_media_link",
                                    "portfolio_link"
                                ],
                                "properties": {
                                    "social_media_link": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://twitter.com/user"
                                    },
                                    "portfolio_link": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://portfolio.com/user"
                                    },
                                    "website_link": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://website.com",
                                        "nullable": true
                                    },
                                    "other_link": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://other.com",
                                        "nullable": true
                                    },
                                    "summary": {
                                        "type": "string",
                                        "example": "Artist specialized in digital art.",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Social media links added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Social media links added successfully."
                                        },
                                        "artistDetail": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid step"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/get-categories": {
            "get": {
                "tags": [
                    "Registration"
                ],
                "summary": "Fetch all categories",
                "operationId": "57cfce92ed141d7357bb852d3598e931",
                "responses": {
                    "200": {
                        "description": "List of categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "categories": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Painting"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/choose-categories": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Choose categories for the user",
                "operationId": "325241fc9fceca51db665c07af5ece56",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categories chosen successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Categories chosen successfully."
                                        },
                                        "artistDetail": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid step"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/add-pickup-location": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Add pickup location for the user",
                "operationId": "f9798fadfab322baff88a81928d1043c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city",
                                    "zone",
                                    "address"
                                ],
                                "properties": {
                                    "city": {
                                        "type": "string",
                                        "example": "Cairo"
                                    },
                                    "zone": {
                                        "type": "string",
                                        "example": "Downtown"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St, Cairo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pickup location added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pickup location added successfully."
                                        },
                                        "artistDetail": {
                                            "type": "object"
                                        },
                                        "address": {
                                            "properties": {
                                                "city": {
                                                    "type": "string",
                                                    "example": "Cairo"
                                                },
                                                "zone": {
                                                    "type": "string",
                                                    "example": "Downtown"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "123 Main St, Cairo"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid step"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use your Bearer token to access secured endpoints.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}